split template code into dedicated files by template type
