+++ apstra/api_design_templates.go
@@ -1,4 +1,4 @@
-// Copyright (c) Juniper Networks, Inc., 2022-2024.
+// Copyright (c) Juniper Networks, Inc., 2022-2025.
 // All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
@@ -7,13 +7,8 @@ package apstra
 import (
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"net/http"
-	"sort"
-	"time"
-
-	"github.com/Juniper/apstra-go-sdk/apstra/compatibility"
 )
 
 const (
@@ -488,159 +483,9 @@ func (o *rawVirtualNetworkPolicy) polish() (*VirtualNetworkPolicy, error) {
 	return &VirtualNetworkPolicy{OverlayControlProtocol: OverlayControlProtocol(ocp)}, err
 }
 
-type AsnAllocationPolicy struct {
-	SpineAsnScheme AsnAllocationScheme
-}
-
-func (o *AsnAllocationPolicy) raw() *rawAsnAllocationPolicy {
-	return &rawAsnAllocationPolicy{SpineAsnScheme: o.SpineAsnScheme.raw()}
-}
-
-type rawAsnAllocationPolicy struct {
-	SpineAsnScheme asnAllocationScheme `json:"spine_asn_scheme"`
-}
-
-func (o *rawAsnAllocationPolicy) polish() (*AsnAllocationPolicy, error) {
-	sas, err := o.SpineAsnScheme.parse()
-	return &AsnAllocationPolicy{SpineAsnScheme: AsnAllocationScheme(sas)}, err
-}
-
 type RackTypeCount struct {
 	RackTypeId ObjectId `json:"rack_type_id"`
 	Count      int      `json:"count"`
-}
-
-type RackBasedTemplateCount struct {
-	RackBasedTemplateId ObjectId `json:"rack_based_template_id"`
-	Count               int      `json:"count"`
-}
-
-type Spine struct {
-	Count                  int
-	LinkPerSuperspineSpeed LogicalDevicePortSpeed
-	LogicalDevice          LogicalDeviceData
-	LinkPerSuperspineCount int
-	Tags                   []DesignTagData
-}
-
-type TemplateElementSpineRequest struct {
-	Count                  int
-	LinkPerSuperspineSpeed LogicalDevicePortSpeed
-	LogicalDevice          ObjectId
-	LinkPerSuperspineCount int
-	Tags                   []ObjectId
-}
-
-func (o *TemplateElementSpineRequest) raw(ctx context.Context, client *Client) (*rawSpine, error) {
-	logicalDevice, err := client.getLogicalDevice(ctx, o.LogicalDevice)
-	if err != nil {
-		return nil, err
-	}
-
-	tags := make([]DesignTagData, len(o.Tags))
-	for i, tagId := range o.Tags {
-		rawTag, err := client.getTag(ctx, tagId)
-		if err != nil {
-			return nil, err
-		}
-		tags[i] = *rawTag.polish().Data
-	}
-
-	return &rawSpine{
-		Count:                  o.Count,
-		LinkPerSuperspineSpeed: o.LinkPerSuperspineSpeed.raw(),
-		LogicalDevice:          *logicalDevice,
-		LinkPerSuperspineCount: o.LinkPerSuperspineCount,
-		Tags:                   tags,
-	}, nil
-}
-
-type rawSpine struct {
-	Count                  int                        `json:"count"`
-	LinkPerSuperspineSpeed *rawLogicalDevicePortSpeed `json:"link_per_superspine_speed"`
-	LogicalDevice          rawLogicalDevice           `json:"logical_device"`
-	LinkPerSuperspineCount int                        `json:"link_per_superspine_count"`
-	Tags                   []DesignTagData            `json:"tags"`
-}
-
-func (o rawSpine) polish() (*Spine, error) {
-	ld, err := o.LogicalDevice.polish()
-
-	var linkPerSuperspineSpeed LogicalDevicePortSpeed
-	if o.LinkPerSuperspineSpeed != nil {
-		linkPerSuperspineSpeed = o.LinkPerSuperspineSpeed.parse()
-	}
-
-	return &Spine{
-		Count:                  o.Count,
-		LinkPerSuperspineSpeed: linkPerSuperspineSpeed,
-		LogicalDevice: LogicalDeviceData{
-			DisplayName: ld.Data.DisplayName,
-			Panels:      ld.Data.Panels,
-		},
-		LinkPerSuperspineCount: o.LinkPerSuperspineCount,
-		Tags:                   o.Tags,
-	}, err
-}
-
-type Superspine struct {
-	PlaneCount         int
-	Tags               []DesignTagData
-	SuperspinePerPlane int
-	LogicalDevice      LogicalDeviceData
-}
-
-type TemplateElementSuperspineRequest struct {
-	PlaneCount         int
-	Tags               []ObjectId
-	SuperspinePerPlane int
-	LogicalDeviceId    ObjectId
-}
-
-func (o *TemplateElementSuperspineRequest) raw(ctx context.Context, client *Client) (*rawSuperspine, error) {
-	tags := make([]DesignTagData, len(o.Tags))
-	for i, tagId := range o.Tags {
-		rawTag, err := client.getTag(ctx, tagId)
-		if err != nil {
-			return nil, err
-		}
-		tags[i] = *rawTag.polish().Data
-	}
-
-	logicalDevice, err := client.getLogicalDevice(ctx, o.LogicalDeviceId)
-	if err != nil {
-		return nil, err
-	}
-
-	return &rawSuperspine{
-		PlaneCount:         o.PlaneCount,
-		Tags:               tags,
-		SuperspinePerPlane: o.SuperspinePerPlane,
-		LogicalDevice:      *logicalDevice,
-	}, nil
-}
-
-type rawSuperspine struct {
-	PlaneCount         int              `json:"plane_count"`
-	Tags               []DesignTagData  `json:"tags"`
-	SuperspinePerPlane int              `json:"superspine_per_plane"`
-	LogicalDevice      rawLogicalDevice `json:"logical_device"`
-}
-
-func (o rawSuperspine) polish() (*Superspine, error) {
-	ld, err := o.LogicalDevice.polish()
-	if err != nil {
-		return nil, err
-	}
-	return &Superspine{
-		PlaneCount:         o.PlaneCount,
-		Tags:               o.Tags,
-		SuperspinePerPlane: o.SuperspinePerPlane,
-		LogicalDevice: LogicalDeviceData{
-			DisplayName: ld.Data.DisplayName,
-			Panels:      ld.Data.Panels,
-		},
-	}, nil
 }
 
 type Template interface {
@@ -698,367 +543,8 @@ func (o *template) polish() (Template, error) {
 	return nil, fmt.Errorf(TemplateTypeUnknown, t)
 }
 
-var _ Template = &TemplateRackBased{}
-
-type TemplateRackBased struct {
-	Id             ObjectId
-	CreatedAt      time.Time
-	LastModifiedAt time.Time
-	templateType   TemplateType
-	Data           *TemplateRackBasedData
-}
-
-func (o *TemplateRackBased) Type() TemplateType {
-	return o.templateType
-}
-
-func (o *TemplateRackBased) ID() ObjectId {
-	return o.Id
-}
-
-func (o *TemplateRackBased) OverlayControlProtocol() OverlayControlProtocol {
-	if o == nil || o.Data == nil {
-		return OverlayControlProtocolNone
-	}
-	return o.Data.VirtualNetworkPolicy.OverlayControlProtocol
-}
-
-type TemplateRackBasedData struct {
-	DisplayName          string
-	AntiAffinityPolicy   *AntiAffinityPolicy
-	VirtualNetworkPolicy VirtualNetworkPolicy
-	AsnAllocationPolicy  AsnAllocationPolicy
-	Capability           TemplateCapability
-	Spine                Spine
-	RackInfo             map[ObjectId]TemplateRackBasedRackInfo
-	DhcpServiceIntent    DhcpServiceIntent
-}
-
-type TemplateRackBasedRackInfo struct {
-	Count        int
-	RackTypeData *RackTypeData
-}
-
 type DhcpServiceIntent struct {
 	Active bool `json:"active"`
-}
-
-type rawTemplateRackBased struct {
-	Id                   ObjectId                `json:"id"`
-	Type                 templateType            `json:"type"`
-	DisplayName          string                  `json:"display_name"`
-	AntiAffinityPolicy   *rawAntiAffinityPolicy  `json:"anti_affinity_policy,omitempty"`
-	CreatedAt            time.Time               `json:"created_at"`
-	LastModifiedAt       time.Time               `json:"last_modified_at"`
-	VirtualNetworkPolicy rawVirtualNetworkPolicy `json:"virtual_network_policy"`
-	AsnAllocationPolicy  rawAsnAllocationPolicy  `json:"asn_allocation_policy"`
-	Capability           templateCapability      `json:"capability,omitempty"`
-	Spine                rawSpine                `json:"spine"`
-	RackTypes            []rawRackType           `json:"rack_types"`
-	RackTypeCounts       []RackTypeCount         `json:"rack_type_counts"`
-	DhcpServiceIntent    DhcpServiceIntent       `json:"dhcp_service_intent"`
-}
-
-func (o rawTemplateRackBased) polish() (*TemplateRackBased, error) {
-	tType, err := o.Type.parse()
-	if err != nil {
-		return nil, err
-	}
-	v, err := o.VirtualNetworkPolicy.polish()
-	if err != nil {
-		return nil, err
-	}
-	a, err := o.AsnAllocationPolicy.polish()
-	if err != nil {
-		return nil, err
-	}
-	c, err := o.Capability.parse()
-	if err != nil {
-		return nil, err
-	}
-	s, err := o.Spine.polish()
-	if err != nil {
-		return nil, err
-	}
-	var aap *AntiAffinityPolicy
-	if o.AntiAffinityPolicy != nil {
-		aap, err = o.AntiAffinityPolicy.polish()
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	if len(o.RackTypes) != len(o.RackTypeCounts) {
-		return nil, fmt.Errorf("template '%s' has %d rack_types and %d rack_type_counts - these should match",
-			o.Id, len(o.RackTypes), len(o.RackTypeCounts))
-	}
-
-	rackTypeInfos := make(map[ObjectId]TemplateRackBasedRackInfo, len(o.RackTypes))
-OUTER:
-	for _, rrt := range o.RackTypes { // loop over raw rack types
-		prt, err := rrt.polish()
-		if err != nil {
-			return nil, err
-		}
-		for _, rtc := range o.RackTypeCounts { // loop over rack type counts looking for matching ID
-			if prt.Id == rtc.RackTypeId {
-				rackTypeInfos[rtc.RackTypeId] = TemplateRackBasedRackInfo{
-					Count:        rtc.Count,
-					RackTypeData: prt.Data,
-				}
-				continue OUTER
-			}
-		}
-		return nil, fmt.Errorf("template contains rack_type '%s' which does not appear among rack_type_counts", rrt.Id)
-	}
-
-	return &TemplateRackBased{
-		Id:             o.Id,
-		templateType:   TemplateType(tType),
-		CreatedAt:      o.CreatedAt,
-		LastModifiedAt: o.LastModifiedAt,
-		Data: &TemplateRackBasedData{
-			DisplayName:          o.DisplayName,
-			AntiAffinityPolicy:   aap,
-			VirtualNetworkPolicy: *v,
-			AsnAllocationPolicy:  *a,
-			Capability:           TemplateCapability(c),
-			Spine:                *s,
-			RackInfo:             rackTypeInfos,
-			DhcpServiceIntent:    o.DhcpServiceIntent,
-		},
-	}, nil
-}
-
-var _ Template = &TemplatePodBased{}
-
-type TemplatePodBased struct {
-	Id             ObjectId
-	templateType   TemplateType
-	CreatedAt      time.Time
-	LastModifiedAt time.Time
-	Data           *TemplatePodBasedData
-}
-
-func (o *TemplatePodBased) Type() TemplateType {
-	return o.templateType
-}
-
-func (o *TemplatePodBased) ID() ObjectId {
-	return o.Id
-}
-
-func (o *TemplatePodBased) OverlayControlProtocol() OverlayControlProtocol {
-	if o == nil || o.Data == nil || len(o.Data.PodInfo) == 0 {
-		return OverlayControlProtocolNone
-	}
-
-	// return the first record
-	for _, v := range o.Data.PodInfo {
-		if v.TemplateRackBasedData != nil {
-			return v.TemplateRackBasedData.VirtualNetworkPolicy.OverlayControlProtocol
-		}
-	}
-
-	return OverlayControlProtocolNone
-}
-
-type TemplatePodBasedData struct {
-	DisplayName        string
-	AntiAffinityPolicy *AntiAffinityPolicy
-	Superspine         Superspine
-	Capability         TemplateCapability
-	PodInfo            map[ObjectId]TemplatePodBasedInfo
-}
-
-type TemplatePodBasedInfo struct {
-	Count                 int
-	TemplateRackBasedData *TemplateRackBasedData
-}
-
-type rawTemplatePodBased struct {
-	Id                      ObjectId                 `json:"id"`
-	Type                    templateType             `json:"type"`
-	DisplayName             string                   `json:"display_name"`
-	AntiAffinityPolicy      *rawAntiAffinityPolicy   `json:"anti_affinity_policy,omitempty"`
-	Superspine              rawSuperspine            `json:"superspine"`
-	CreatedAt               time.Time                `json:"created_at"`
-	LastModifiedAt          time.Time                `json:"last_modified_at"`
-	Capability              templateCapability       `json:"capability,omitempty"`
-	RackBasedTemplates      []rawTemplateRackBased   `json:"rack_based_templates"`
-	RackBasedTemplateCounts []RackBasedTemplateCount `json:"rack_based_template_counts"`
-}
-
-func (o rawTemplatePodBased) polish() (*TemplatePodBased, error) {
-	superspine, err := o.Superspine.polish()
-	if err != nil {
-		return nil, err
-	}
-	capability, err := o.Capability.parse()
-	if err != nil {
-		return nil, err
-	}
-	var _, rbt []TemplateRackBased
-	for _, rrbt := range o.RackBasedTemplates {
-		if rrbt.Type == templateTypeNone {
-			// because sometimes Apstra doesn't fill this in, but we know based on context
-			rrbt.Type = templateTypeRackBased
-		}
-		polished, err := rrbt.polish()
-		if err != nil {
-			return nil, err
-		}
-		rbt = append(rbt, *polished)
-	}
-	tType, err := o.Type.parse()
-	if err != nil {
-		return nil, err
-	}
-	var aap *AntiAffinityPolicy
-	if o.AntiAffinityPolicy != nil {
-		aap, err = o.AntiAffinityPolicy.polish()
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	if len(o.RackBasedTemplates) != len(o.RackBasedTemplateCounts) {
-		return nil, fmt.Errorf("template '%s' has %d rack_based_templates and %d rack_based_template_counts - "+
-			"these should match", o.Id, len(o.RackBasedTemplates), len(o.RackBasedTemplateCounts))
-	}
-
-	podTypeInfos := make(map[ObjectId]TemplatePodBasedInfo, len(o.RackBasedTemplates))
-OUTER:
-	for _, rrbt := range o.RackBasedTemplates {
-		prbt, err := rrbt.polish()
-		if err != nil {
-			return nil, err
-		}
-		for _, rbtc := range o.RackBasedTemplateCounts { // loop over rack based template counts looking for matching ID
-			if prbt.Id == rbtc.RackBasedTemplateId {
-				podTypeInfos[rbtc.RackBasedTemplateId] = TemplatePodBasedInfo{
-					Count:                 rbtc.Count,
-					TemplateRackBasedData: prbt.Data,
-				}
-				continue OUTER
-			}
-		}
-		return nil, fmt.Errorf("template contains rack_based_template '%s' which does not appear among rack_based_tempalte_counts", rrbt.Id)
-	}
-
-	return &TemplatePodBased{
-		Id:             o.Id,
-		templateType:   TemplateType(tType),
-		CreatedAt:      o.CreatedAt,
-		LastModifiedAt: o.LastModifiedAt,
-		Data: &TemplatePodBasedData{
-			DisplayName:        o.DisplayName,
-			AntiAffinityPolicy: aap,
-			Superspine:         *superspine,
-			Capability:         TemplateCapability(capability),
-			PodInfo:            podTypeInfos,
-		},
-	}, nil
-}
-
-var _ Template = &TemplatePodBased{}
-
-type TemplateL3Collapsed struct {
-	Id             ObjectId
-	templateType   TemplateType
-	CreatedAt      time.Time
-	LastModifiedAt time.Time
-	Data           *TemplateL3CollapsedData
-}
-
-func (o *TemplateL3Collapsed) Type() TemplateType {
-	return o.templateType
-}
-
-func (o *TemplateL3Collapsed) ID() ObjectId {
-	return o.Id
-}
-
-func (o *TemplateL3Collapsed) OverlayControlProtocol() OverlayControlProtocol {
-	if o == nil || o.Data == nil {
-		return OverlayControlProtocolNone
-	}
-	return o.Data.VirtualNetworkPolicy.OverlayControlProtocol
-}
-
-type TemplateL3CollapsedData struct {
-	DisplayName          string
-	AntiAffinityPolicy   *AntiAffinityPolicy
-	RackTypes            []RackType
-	Capability           TemplateCapability
-	MeshLinkSpeed        LogicalDevicePortSpeed
-	VirtualNetworkPolicy VirtualNetworkPolicy
-	MeshLinkCount        int
-	RackTypeCounts       []RackTypeCount
-	DhcpServiceIntent    DhcpServiceIntent
-}
-
-type rawTemplateL3Collapsed struct {
-	Id                   ObjectId                   `json:"id"`
-	Type                 templateType               `json:"type"`
-	DisplayName          string                     `json:"display_name"`
-	AntiAffinityPolicy   *rawAntiAffinityPolicy     `json:"anti_affinity_policy,omitempty"`
-	CreatedAt            time.Time                  `json:"created_at"`
-	LastModifiedAt       time.Time                  `json:"last_modified_at"`
-	RackTypes            []rawRackType              `json:"rack_types"`
-	Capability           templateCapability         `json:"capability"`
-	MeshLinkSpeed        *rawLogicalDevicePortSpeed `json:"mesh_link_speed"`
-	VirtualNetworkPolicy rawVirtualNetworkPolicy    `json:"virtual_network_policy"`
-	MeshLinkCount        int                        `json:"mesh_link_count"`
-	RackTypeCounts       []RackTypeCount            `json:"rack_type_counts"`
-	DhcpServiceIntent    DhcpServiceIntent          `json:"dhcp_service_intent"`
-}
-
-func (o rawTemplateL3Collapsed) polish() (*TemplateL3Collapsed, error) {
-	var prt []RackType
-	for _, rrt := range o.RackTypes {
-		polished, err := rrt.polish()
-		if err != nil {
-			return nil, err
-		}
-		prt = append(prt, *polished)
-	}
-	capability, err := o.Capability.parse()
-	if err != nil {
-		return nil, err
-	}
-	vnp, err := o.VirtualNetworkPolicy.polish()
-	if err != nil {
-		return nil, err
-	}
-	tType, err := o.Type.parse()
-	if err != nil {
-		return nil, err
-	}
-	var aap *AntiAffinityPolicy
-	if o.AntiAffinityPolicy != nil {
-		aap, err = o.AntiAffinityPolicy.polish()
-		if err != nil {
-			return nil, err
-		}
-	}
-	return &TemplateL3Collapsed{
-		Id:             o.Id,
-		templateType:   TemplateType(tType),
-		CreatedAt:      o.CreatedAt,
-		LastModifiedAt: o.LastModifiedAt,
-		Data: &TemplateL3CollapsedData{
-			DisplayName:          o.DisplayName,
-			AntiAffinityPolicy:   aap,
-			RackTypes:            prt,
-			Capability:           TemplateCapability(capability),
-			MeshLinkSpeed:        o.MeshLinkSpeed.parse(),
-			VirtualNetworkPolicy: *vnp,
-			MeshLinkCount:        o.MeshLinkCount,
-			RackTypeCounts:       o.RackTypeCounts,
-			DhcpServiceIntent:    o.DhcpServiceIntent,
-		},
-	}, nil
 }
 
 func (o *Client) listAllTemplateIds(ctx context.Context) ([]ObjectId, error) {
@@ -1114,118 +600,6 @@ func (o *Client) getAllTemplates(ctx context.Context) ([]template, error) {
 	return result, nil
 }
 
-func (o *Client) getRackBasedTemplate(ctx context.Context, id ObjectId) (*rawTemplateRackBased, error) {
-	rawTemplate, err := o.getTemplate(ctx, id)
-	if err != nil {
-		return nil, err
-	}
-
-	tType, err := rawTemplate.templateType()
-	if err != nil {
-		return nil, err
-	}
-
-	if tType != templateTypeRackBased {
-		return nil, ClientErr{
-			errType: ErrWrongType,
-			err:     fmt.Errorf("template '%s' is of type '%s', not '%s'", id, tType, templateTypeRackBased),
-		}
-	}
-
-	template := &rawTemplateRackBased{}
-	return template, json.Unmarshal(rawTemplate, template)
-}
-
-func (o *Client) getAllRackBasedTemplates(ctx context.Context) ([]rawTemplateRackBased, error) {
-	templates, err := o.getAllTemplates(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	var result []rawTemplateRackBased
-	for _, t := range templates {
-		tType, err := t.templateType()
-		if err != nil {
-			return nil, err
-		}
-		if tType != templateTypeRackBased {
-			continue
-		}
-		var raw rawTemplateRackBased
-		err = json.Unmarshal(t, &raw)
-		if err != nil {
-			return nil, err
-		}
-		result = append(result, raw)
-	}
-
-	return result, nil
-}
-
-func (o *Client) getPodBasedTemplate(ctx context.Context, id ObjectId) (*rawTemplatePodBased, error) {
-	rawTemplate, err := o.getTemplate(ctx, id)
-	if err != nil {
-		return nil, err
-	}
-
-	tType, err := rawTemplate.templateType()
-	if err != nil {
-		return nil, err
-	}
-
-	if tType != templateTypePodBased {
-		return nil, ClientErr{
-			errType: ErrWrongType,
-			err:     fmt.Errorf("template '%s' is of type '%s', not '%s'", id, tType, templateTypePodBased),
-		}
-	}
-
-	result := &rawTemplatePodBased{}
-	err = json.Unmarshal(rawTemplate, result)
-	if err != nil {
-		return nil, fmt.Errorf("error unmarshaling raw pod-based template - %w", err)
-	}
-
-	// force 'type' field of included rack-based templates to "rack_based" b/c Apstra rejects empty string.
-	for i, rbt := range result.RackBasedTemplates {
-		switch rbt.Type {
-		case "":
-			result.RackBasedTemplates[i].Type = templateTypeRackBased
-		case templateTypeRackBased: // fallthrough
-		default:
-			return nil, fmt.Errorf("rack-based template '%s' within pod-based template '%s' claims to be type '%s', expected '%s'",
-				rbt.DisplayName, result.Id, rbt.Type, templateTypeRackBased)
-		}
-	}
-	return result, nil
-}
-
-func (o *Client) getAllPodBasedTemplates(ctx context.Context) ([]rawTemplatePodBased, error) {
-	templates, err := o.getAllTemplates(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	var result []rawTemplatePodBased
-	for _, t := range templates {
-		tType, err := t.templateType()
-		if err != nil {
-			return nil, err
-		}
-		if tType != templateTypePodBased {
-			continue
-		}
-		var raw rawTemplatePodBased
-		err = json.Unmarshal(t, &raw)
-		if err != nil {
-			return nil, err
-		}
-		result = append(result, raw)
-	}
-
-	return result, nil
-}
-
 func (o *Client) getTemplateByTypeAndName(ctx context.Context, desiredType templateType, desiredName string) (*template, error) {
 	templates, err := o.getAllTemplates(ctx)
 	if err != nil {
@@ -1269,362 +643,6 @@ func (o *Client) getTemplateByTypeAndName(ctx context.Context, desiredType templ
 	}
 
 	return found, nil
-}
-
-func (o *Client) getL3CollapsedTemplate(ctx context.Context, id ObjectId) (*rawTemplateL3Collapsed, error) {
-	rawTemplate, err := o.getTemplate(ctx, id)
-	if err != nil {
-		return nil, err
-	}
-
-	tType, err := rawTemplate.templateType()
-	if err != nil {
-		return nil, err
-	}
-
-	if tType != templateTypeL3Collapsed {
-		return nil, ClientErr{
-			errType: ErrWrongType,
-			err:     fmt.Errorf("template '%s' is of type '%s', not '%s'", id, tType, templateTypeL3Collapsed),
-		}
-	}
-
-	template := &rawTemplateL3Collapsed{}
-	return template, json.Unmarshal(rawTemplate, template)
-}
-
-func (o *Client) getAllL3CollapsedTemplates(ctx context.Context) ([]rawTemplateL3Collapsed, error) {
-	templates, err := o.getAllTemplates(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	var result []rawTemplateL3Collapsed
-	for _, t := range templates {
-		tType, err := t.templateType()
-		if err != nil {
-			return nil, err
-		}
-		if tType != templateTypeL3Collapsed {
-			continue
-		}
-		var raw rawTemplateL3Collapsed
-		err = json.Unmarshal(t, &raw)
-		if err != nil {
-			return nil, err
-		}
-		result = append(result, raw)
-	}
-
-	return result, nil
-}
-
-type CreateRackBasedTemplateRequest struct {
-	DisplayName          string
-	Spine                *TemplateElementSpineRequest
-	RackInfos            map[ObjectId]TemplateRackBasedRackInfo
-	DhcpServiceIntent    *DhcpServiceIntent
-	AntiAffinityPolicy   *AntiAffinityPolicy
-	AsnAllocationPolicy  *AsnAllocationPolicy
-	VirtualNetworkPolicy *VirtualNetworkPolicy
-}
-
-func (o *CreateRackBasedTemplateRequest) raw(ctx context.Context, client *Client) (*rawCreateRackBasedTemplateRequest, error) {
-	rackTypes := make([]rawRackType, len(o.RackInfos))
-	rackTypeCounts := make([]RackTypeCount, len(o.RackInfos))
-	var i int
-	for k, ri := range o.RackInfos {
-		if ri.RackTypeData != nil {
-			return nil, fmt.Errorf("the RackTypeData field must be nil when creating a rack-based template")
-		}
-
-		// grab the rack type from the API using the caller's map key (ObjectId) and stash it in rackTypes
-		rt, err := client.getRackType(ctx, k)
-		if err != nil {
-			return nil, err
-		}
-		rackTypes[i] = *rt
-
-		// prep the RackTypeCount object using the caller's map key (ObjectId) as
-		// the link between the rawRackType data copy and the RackTypeCount
-		rackTypeCounts[i].RackTypeId = k
-		rackTypeCounts[i].Count = ri.Count
-		i++
-	}
-
-	switch {
-	case o.Spine == nil:
-		return nil, errors.New("spine cannot be <nil> when creating a rack-based template")
-	case o.AntiAffinityPolicy == nil && compatibility.TemplateRequestRequiresAntiAffinityPolicy.Check(client.apiVersion):
-		return nil, fmt.Errorf("anti-affinity policy cannot be <nil> when creating a rack-based template with Apstra %s", compatibility.TemplateRequestRequiresAntiAffinityPolicy)
-	case o.AsnAllocationPolicy == nil:
-		return nil, errors.New("asn allocation policy cannot be <nil> when creating a rack-based template")
-	case o.VirtualNetworkPolicy == nil:
-		return nil, errors.New("virtual network policy cannot be <nil> when creating a rack-based template")
-	}
-
-	var err error
-	var dhcpServiceIntent DhcpServiceIntent
-	if o.DhcpServiceIntent != nil {
-		dhcpServiceIntent = *o.DhcpServiceIntent
-	}
-
-	spine, err := o.Spine.raw(ctx, client)
-	if err != nil {
-		return nil, err
-	}
-
-	var antiAffinityPolicy *rawAntiAffinityPolicy
-	if o.AntiAffinityPolicy != nil {
-		antiAffinityPolicy = o.AntiAffinityPolicy.raw()
-	}
-
-	asnAllocationPolicy := o.AsnAllocationPolicy.raw()
-
-	virtualNetworkPolicy := o.VirtualNetworkPolicy.raw()
-
-	return &rawCreateRackBasedTemplateRequest{
-		Type:                 templateTypeRackBased,
-		DisplayName:          o.DisplayName,
-		Spine:                *spine,
-		RackTypes:            rackTypes,
-		RackTypeCounts:       rackTypeCounts,
-		DhcpServiceIntent:    dhcpServiceIntent,
-		AntiAffinityPolicy:   antiAffinityPolicy,
-		AsnAllocationPolicy:  *asnAllocationPolicy,
-		VirtualNetworkPolicy: *virtualNetworkPolicy,
-	}, nil
-}
-
-type rawCreateRackBasedTemplateRequest struct {
-	Type                 templateType            `json:"type"`
-	DisplayName          string                  `json:"display_name"`
-	Spine                rawSpine                `json:"spine"`
-	RackTypes            []rawRackType           `json:"rack_types"`
-	RackTypeCounts       []RackTypeCount         `json:"rack_type_counts"`
-	DhcpServiceIntent    DhcpServiceIntent       `json:"dhcp_service_intent"`
-	AntiAffinityPolicy   *rawAntiAffinityPolicy  `json:"anti_affinity_policy,omitempty"`
-	AsnAllocationPolicy  rawAsnAllocationPolicy  `json:"asn_allocation_policy"`
-	VirtualNetworkPolicy rawVirtualNetworkPolicy `json:"virtual_network_policy"`
-}
-
-func (o *Client) createRackBasedTemplate(ctx context.Context, in *rawCreateRackBasedTemplateRequest) (ObjectId, error) {
-	response := &objectIdResponse{}
-	err := o.talkToApstra(ctx, &talkToApstraIn{
-		method:      http.MethodPost,
-		urlStr:      apiUrlDesignTemplates,
-		apiInput:    in,
-		apiResponse: response,
-	})
-	if err != nil {
-		return "", convertTtaeToAceWherePossible(err)
-	}
-	return response.Id, nil
-}
-
-func (o *Client) updateRackBasedTemplate(ctx context.Context, id ObjectId, in *CreateRackBasedTemplateRequest) error {
-	raw, err := in.raw(ctx, o)
-	if err != nil {
-		return err
-	}
-	err = o.talkToApstra(ctx, &talkToApstraIn{
-		method:   http.MethodPut,
-		urlStr:   fmt.Sprintf(apiUrlDesignTemplateById, id),
-		apiInput: raw,
-	})
-	if err != nil {
-		return convertTtaeToAceWherePossible(err)
-	}
-	return nil
-}
-
-type CreatePodBasedTemplateRequest struct {
-	DisplayName        string
-	Superspine         *TemplateElementSuperspineRequest
-	PodInfos           map[ObjectId]TemplatePodBasedInfo
-	AntiAffinityPolicy *AntiAffinityPolicy
-}
-
-func (o *CreatePodBasedTemplateRequest) raw(ctx context.Context, client *Client) (*rawCreatePodBasedTemplateRequest, error) {
-	templatesRackBased := make([]rawTemplateRackBased, len(o.PodInfos))
-	rackBasedTemplatesCounts := make([]RackBasedTemplateCount, len(o.PodInfos))
-	var i int
-	for k, pi := range o.PodInfos {
-		if pi.TemplateRackBasedData != nil {
-			return nil, fmt.Errorf("the TemplateRackBasedData (pod info) field must be nil when creating a pod-based template")
-		}
-
-		// grab the rack-based template (pod) from the API using the caller's map key (ObjectId) and stash it in templatesRackBased
-		rbt, err := client.getRackBasedTemplate(ctx, k)
-		if err != nil {
-			return nil, err
-		}
-		templatesRackBased[i] = *rbt
-
-		// prep the RackBasedTemplateCount object using the caller's map key (ObjectId) as
-		// the link between the rawTemplateRackBased data copy and the RackBasedTemplateCount
-		rackBasedTemplatesCounts[i].RackBasedTemplateId = k
-		rackBasedTemplatesCounts[i].Count = pi.Count
-		i++
-	}
-
-	sort.Slice(templatesRackBased, func(i, j int) bool {
-		return templatesRackBased[i].DisplayName < templatesRackBased[j].DisplayName
-	})
-
-	switch {
-	case o.Superspine == nil:
-		return nil, errors.New("super spine cannot be <nil> when creating a pod-based template")
-	case o.AntiAffinityPolicy == nil && compatibility.TemplateRequestRequiresAntiAffinityPolicy.Check(client.apiVersion):
-		return nil, fmt.Errorf("anti-affinity policy cannot be <nil> when creating a pod-based template with Apstra %s", compatibility.TemplateRequestRequiresAntiAffinityPolicy)
-	}
-
-	var err error
-	var superspine *rawSuperspine
-	if o.Superspine != nil {
-		superspine, err = o.Superspine.raw(ctx, client)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	var antiAffinityPolicy *rawAntiAffinityPolicy
-	if o.AntiAffinityPolicy != nil {
-		antiAffinityPolicy = o.AntiAffinityPolicy.raw()
-	}
-
-	return &rawCreatePodBasedTemplateRequest{
-		Type:                    templateTypePodBased,
-		DisplayName:             o.DisplayName,
-		Superspine:              *superspine,
-		RackBasedTemplates:      templatesRackBased,
-		RackBasedTemplateCounts: rackBasedTemplatesCounts,
-		AntiAffinityPolicy:      antiAffinityPolicy,
-	}, nil
-}
-
-type rawCreatePodBasedTemplateRequest struct {
-	Type                    templateType             `json:"type"`
-	DisplayName             string                   `json:"display_name"`
-	Superspine              rawSuperspine            `json:"superspine"`
-	RackBasedTemplates      []rawTemplateRackBased   `json:"rack_based_templates"`
-	RackBasedTemplateCounts []RackBasedTemplateCount `json:"rack_based_template_counts"`
-	AntiAffinityPolicy      *rawAntiAffinityPolicy   `json:"anti_affinity_policy,omitempty"`
-}
-
-func (o *Client) createPodBasedTemplate(ctx context.Context, in *rawCreatePodBasedTemplateRequest) (ObjectId, error) {
-	response := &objectIdResponse{}
-	err := o.talkToApstra(ctx, &talkToApstraIn{
-		method:      http.MethodPost,
-		urlStr:      apiUrlDesignTemplates,
-		apiInput:    in,
-		apiResponse: response,
-	})
-	if err != nil {
-		return "", err
-	}
-
-	return response.Id, nil
-}
-
-func (o *Client) updatePodBasedTemplate(ctx context.Context, id ObjectId, in *CreatePodBasedTemplateRequest) error {
-	apiInput, err := in.raw(ctx, o)
-	if err != nil {
-		return err
-	}
-	err = o.talkToApstra(ctx, &talkToApstraIn{
-		method:   http.MethodPut,
-		urlStr:   fmt.Sprintf(apiUrlDesignTemplateById, id),
-		apiInput: apiInput,
-	})
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-type CreateL3CollapsedTemplateRequest struct {
-	DisplayName          string                 `json:"display_name"`
-	MeshLinkCount        int                    `json:"mesh_link_count"`
-	MeshLinkSpeed        LogicalDevicePortSpeed `json:"mesh_link_speed"`
-	RackTypeIds          []ObjectId             `json:"rack_types"`
-	RackTypeCounts       []RackTypeCount        `json:"rack_type_counts"`
-	DhcpServiceIntent    DhcpServiceIntent      `json:"dhcp_service_intent"`
-	AntiAffinityPolicy   *AntiAffinityPolicy    `json:"anti_affinity_policy,omitempty"`
-	VirtualNetworkPolicy VirtualNetworkPolicy   `json:"virtual_network_policy"`
-}
-
-func (o *CreateL3CollapsedTemplateRequest) raw(ctx context.Context, client *Client) (*rawCreateL3CollapsedTemplateRequest, error) {
-	rackTypes := make([]rawRackType, len(o.RackTypeIds))
-	for i, id := range o.RackTypeIds {
-		rt, err := client.getRackType(ctx, id)
-		if err != nil {
-			return nil, err
-		}
-		rackTypes[i] = *rt
-	}
-
-	var antiAffinityPolicy *rawAntiAffinityPolicy
-	if o.AntiAffinityPolicy != nil {
-		antiAffinityPolicy = o.AntiAffinityPolicy.raw()
-	}
-
-	return &rawCreateL3CollapsedTemplateRequest{
-		Type:                 templateTypeL3Collapsed,
-		DisplayName:          o.DisplayName,
-		MeshLinkCount:        o.MeshLinkCount,
-		MeshLinkSpeed:        *o.MeshLinkSpeed.raw(),
-		RackTypes:            rackTypes,
-		RackTypeCounts:       o.RackTypeCounts,
-		DhcpServiceIntent:    o.DhcpServiceIntent,
-		AntiAffinityPolicy:   antiAffinityPolicy,
-		VirtualNetworkPolicy: *o.VirtualNetworkPolicy.raw(),
-	}, nil
-}
-
-type rawCreateL3CollapsedTemplateRequest struct {
-	Type                 templateType              `json:"type"`
-	DisplayName          string                    `json:"display_name"`
-	MeshLinkCount        int                       `json:"mesh_link_count"`
-	MeshLinkSpeed        rawLogicalDevicePortSpeed `json:"mesh_link_speed"`
-	RackTypes            []rawRackType             `json:"rack_types"`
-	RackTypeCounts       []RackTypeCount           `json:"rack_type_counts"`
-	DhcpServiceIntent    DhcpServiceIntent         `json:"dhcp_service_intent"`
-	AntiAffinityPolicy   *rawAntiAffinityPolicy    `json:"anti_affinity_policy,omitempty"`
-	VirtualNetworkPolicy rawVirtualNetworkPolicy   `json:"virtual_network_policy"`
-}
-
-func (o *Client) createL3CollapsedTemplate(ctx context.Context, in *rawCreateL3CollapsedTemplateRequest) (ObjectId, error) {
-	response := &objectIdResponse{}
-	err := o.talkToApstra(ctx, &talkToApstraIn{
-		method:      http.MethodPost,
-		urlStr:      apiUrlDesignTemplates,
-		apiInput:    in,
-		apiResponse: response,
-	})
-	if err != nil {
-		return "", err
-	}
-
-	return response.Id, nil
-}
-
-func (o *Client) updateL3CollapsedTemplate(ctx context.Context, id ObjectId, in *CreateL3CollapsedTemplateRequest) error {
-	apiInput, err := in.raw(ctx, o)
-	if err != nil {
-		return err
-	}
-	err = o.talkToApstra(ctx, &talkToApstraIn{
-		method:   http.MethodPut,
-		urlStr:   fmt.Sprintf(apiUrlDesignTemplateById, id),
-		apiInput: apiInput,
-	})
-	if err != nil {
-		return err
-	}
-
-	return nil
 }
 
 func (o *Client) deleteTemplate(ctx context.Context, id ObjectId) error {
